// examples/src/db.d.ts
/**
 * This file was generated by kysely-codegen.
 * Please do not edit it manually.
 */

import type { ColumnType } from "kysely";

export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
  ? ColumnType<S, I | undefined, U>
  : ColumnType<T, T | undefined, T>;

export type Timestamp = ColumnType<Date, Date | string, Date | string>;

export interface Comments {
  authorName: string;
  commentText: string;
  createdAt: Generated<Timestamp | null>;
  id: Generated<number>;
  postId: number | null;
}

export interface Posts {
  content: string;
  id: Generated<number>;
  publishedAt: Generated<Timestamp | null>;
  title: string;
  updatedAt: Generated<Timestamp | null>;
}

export interface DB {
  comments: Comments;
  posts: Posts;
}